% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seagull.R
\name{flock}
\alias{flock}
\title{Low-level flock object}
\usage{
flock(filename, method = "fcntl")
}
\arguments{
\item{filename}{Name of file to lock.  \code{NULL} is a fake lock;
acquire always succeeds.}

\item{method}{Method to use}
}
\description{
Low-level flock object.  Use this if you need more flexibility
than \code{\link{with_flock}}, but understand that if you get it
wrong you can cause deadlocks.
}
\section{Methods}{


\describe{
\item{\code{acquire}}{
  Attempt to acquire a lock on the file.  For details on arguments see the main argument section above.

  \emph{Usage:}
  \code{acquire(delay = 0.01, max_delay = 0.1, timeout = Inf, error = TRUE,
      verbose = FALSE)}

  \emph{Arguments:}
  \itemize{
    \item{\code{delay}:   Initial delay
    }

    \item{\code{max_delay}:   Maximum delay between attempts
    }

    \item{\code{timeout}:   Maximum time to wait
    }

    \item{\code{error}:   Throw an error on failure?
    }

    \item{\code{verbose}:   Print information while attempting to acquire lock?
    }
  }
}
\item{\code{release}}{
  Release the lock.

  \emph{Usage:}
  \code{release()}
}
}
}

