% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\docType{data}
\name{phenotypes}
\alias{phenotypes}
\title{Simulated phenotypes}
\format{A matrix with of 1000 rows and 3 columns. One row per individual. One
column per trait.}
\usage{
phenotypes
}
\description{
A matrix consisting of 1000 rows and 3 columns. Each row
corresponds to a different individual. Each column corresponds to a different
trait. The different traits were simulated to be uncorrelated to one another.
The trait in the first, second, and third column have a heritability equal to
0.1, 0.3, and 0.5, respectively.
}
\author{
Jan Klosa \email{klosa@fbn-dummerstorf.de}
}
\keyword{datasets}
