% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seagull.R
\name{flock}
\alias{flock}
\title{Low-level flock object}
\usage{
flock(filename, method = "fcntl")
}
\arguments{
\item{filename}{Name of file to lock}

\item{method}{Method to use}
}
\description{
Low-level flock object.  Use this if you need more flexibility
than \code{\link{with_flock}}, but understand that if you get it
wrong you can cause deadlocks.
}

